// src/app/modules/user/user.interface.ts
import { Model, Types } from 'mongoose';
import { IStudent } from '../student/student.interface';
import { IFaculty } from '../faculty/faculty.interface';

export type IUser = {
  id: string;
  role: string;
  password: string;
  student?: Types.ObjectId | IStudent;
  faculty?: Types.ObjectId | IFaculty; //made by self
  // admin?: Types.ObjectId | IAdmin; made by self
};

export type UserModel = Model<IUser, Record<string, unknown>>;